+++ maintenance-bungee/src/main/java/eu/kennytv/maintenance/bungee/listener/ProxyPingListener.java
@@ -24,7 +24,6 @@ import net.md_5.bungee.api.ServerPing;
 import net.md_5.bungee.api.event.ProxyPingEvent;
 import net.md_5.bungee.api.plugin.Listener;
 import net.md_5.bungee.event.EventHandler;
-import net.md_5.bungee.event.EventPriority;
 
 public final class ProxyPingListener implements Listener {
     private final MaintenanceBungeePlugin plugin;