+++ maintenance-core/src/main/java/eu/kennytv/maintenance/core/Settings.java
@@ -358,6 +358,7 @@ public class Settings implements ISettings {
                 break;
             }
         }
+
         if (uuid == null) return false;
 
         whitelistedPlayers.remove(uuid);