+++ maintenance-core/src/main/java/eu/kennytv/maintenance/core/listener/JoinListenerBase.java
@@ -67,6 +67,7 @@ public abstract class JoinListenerBase {
         if (!sender.hasPermission("maintenance.admin") || notifiedPlayers.contains(sender.getUuid())) return;
         plugin.async(() -> {
             if (!plugin.updateAvailable()) return;
+
             notifiedPlayers.add(sender.getUuid());
             sender.sendMessage(plugin.getPrefix() + "§cThere is a newer version available: §aVersion " + plugin.getNewestVersion() + "§c, you're on §a" + plugin.getVersion());
             plugin.sendUpdateNotification(sender);